NewRealese

+++ app/src/main/java/com/capstone_bangkit/nutrise/data/AnalysisResult.kt
@@ -2,6 +2,6 @@ package com.capstone_bangkit.nutrise.data
 
 data class AnalysisResult(
     val food: String,
-    val calories: Int,
-    val protein: Int
+    val calories: Double,
+    val protein: Double
 )
